Rewrite YAML parsing to avoid YamlBeans reflection.
Recent versions of YamlBeans require that a writable property `String foo` must have all of `private String foo;`, `public String getFoo()` and `public void setFoo(String foo)`. This is so that properties can be enumerated by looking at the fields. We could in principle change all our `config` beans to conform to this, but it's easier just to bypass YamlBeans reflection. We can ask YamlBeans to parse into the generic `Map` type and then figure out ourselves how to construct the required objects and call the required setters.

PiperOrigin-RevId: 562878144
Change-Id: I89f7090dc37533f072d1c01a850661f949b75da3