Handle reflective access failure for default charset in newer Java versions.
On Java versions 26 and above, reflective modification of the default charset is blocked. Instead of throwing a `LinkageError`, log a warning. The tests for `FileEncodingSetter` are updated to only run on Java versions 25 and below, where this reflective operation is still permitted.

Also Add support for EE11 in ApplicationTest as a test parameter, adds Java version constraints for different EE versions, and updates expected staging outputs for EE11, including JSP-related libraries and context parameters.

PiperOrigin-RevId: 811157886
Change-Id: I05f03b092cae94c866f01f85d7472b593ec53008